refactor: creates class for apply decorations
